+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - configuration for "TBR out of control !!!".
+- configuration for "Azimut".
 
 ## [1.10.0] - 2024-12-31
 ### Added